+++ src/Opaleye.hs
@@ -37,6 +37,7 @@ module Opaleye ( module Opaleye.Adaptors
                , module Opaleye.ToFields
                , module Opaleye.Values
                , module Opaleye.With
+               , module Opaleye.Window
                ) where
 
 import Opaleye.Adaptors
@@ -62,4 +63,5 @@ import Opaleye.SqlTypes
 import Opaleye.Table
 import Opaleye.ToFields
 import Opaleye.Values
+import Opaleye.Window
 import Opaleye.With